backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
websockets>=14.0
sqlalchemy>=2.0
aiosqlite>=0.20
langgraph>=0.4
langchain-core>=0.3
google-generativeai>=0.8
pydantic>=2.0
python-dotenv>=1.0
